Moving aggregation builder method to root level aggregation class
